Remove misleading info about rooms being deleted after last user leaves
Synapse does not do this. That is an open request but it has not been implemented. https://github.com/matrix-org/synapse/issues/4720

I sent a message in May,https://matrix.to/#/!tIWQjETAJquswcgSTQ:tu-dresden.de/$nbbHykJfc8OJBeQ2R0hbvf3mgAtczsGSTXPidH5nzkc?via=tu-dresden.de&via=matrix.org&via=ungleich.ch, asking  if this was a custom feature implemented by your university or if this sentence was a result of misunderstanding how Synapse works. I never got a response so I am making this PR to remove the misleading information. If this is a custom feature then please clarify that in the documentation.